Implement requested feedback
